requirements.txt
pandas
tabula-py
SQLAlchemy
bs4
requests
selenium
splinter>=0.7.3
install-jdk